Add files via upload
Universal Health Coverage (UHC) means that all people and communities receive the quality of health services they need without facing financial hardship. This is pivotal to the mission of the Ministry of Health and Social Services (MoHSS), which is to provide integrated, affordable, accessible, equitable, quality health and social services that are responsive to the needs of the population. As a Member State of the World Health Organisation (WHO) and having subscribed to the Sustainable Development Goal (SDG) 3: Ensure Healthy Lives and Promote Wellbeing for All at All Ages, Namibia has embraced the global effort to work towards achieving Universal Health Coverage (UHC).